# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "TPVV",
    "description" : "TPVV Ingeniería Web",
    "contact" : {
      "email" : "lac56@alu.ua.es"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/lac56-alu/TPVV/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://localhost:2000",
    "description" : "Desarrollo servidor basado en TPVV IW"
  }, {
    "url" : "",
    "description" : "Servidor"
  } ],
...
  "paths" : {
    "/tpvv/registro" : {
      "put" : {
        "tags" : [ "Seguridad" ],
        "summary" : "Se da de alta a un nuevo usuario en nuestro sistema",
        "description" : "Podremos aregrar nuevos usuario/clientes a nuestro sistema a través de un sencillo formulario de registro.",
        "operationId" : "registro",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Registro"
        },
        "responses" : {
          "201" : {
            "description" : "Registro realizado correctamente",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
...
    "/tpvv/token" : {
      "get" : {
        "tags" : [ "Seguridad" ],
        "summary" : "Se obtiene el token único del usuario",
        "description" : "Para realizar la aplicación tendremos dos tipos de autorización, una apiKey que nos dará la autorización para hacer las peticiones de la API y un AuthToken único para cada uno de los usuarios de nuestro TPVV.",
        "operationId" : "getToken",
        "responses" : {
          "201" : {
            "description" : "Nos devolverá el Token de ese usuario.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          },
          "403" : {
            "description" : "El cliente no posee los permisos necesarios para cierto contenido, por lo que el servidor está rechazando otorgar una respuesta apropiada.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Usuario" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "nombre" : {
            "type" : "string",
            "example" : "Luis Alfonso"
          },
          "email" : {
            "type" : "string",
            "example" : "luis@gmail.com"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "example" : "aijsfbhausfbas"
          },
          "token" : {
            "type" : "string",
            "example" : "i989nasiudfbas54asd5as4da5s1d"
          },
          "nombreEmpresa" : {
            "type" : "string",
            "example" : "A reason"
          },
          "tipoUsuario" : {
            "type" : "string",
            "example" : "normal"
          }
        },
        "example" : {
          "password" : "aijsfbhausfbas",
          "tipoUsuario" : "normal",
          "id" : 1,
          "nombre" : "Luis Alfonso",
          "email" : "luis@gmail.com",
          "nombreEmpresa" : "A reason",
          "token" : "i989nasiudfbas54asd5as4da5s1d"
        }
      },
...
      "inline_response_400" : {
        "type" : "object",
        "properties" : {
          "typeError" : {
            "type" : "string",
            "example" : "ERROR_PARAMETERS"
          },
          "messageError" : {
            "type" : "string",
            "example" : "Los parámetros introducidos son erroneos."
          }
        }
      },
      "inline_response_201" : {
        "type" : "object",
        "properties" : {
          "Token" : {
            "type" : "string",
            "example" : "jbHBSFBUIAKONaiaoizfcja0f09kasndjan"
          }
        },
        "example" : {
          "Token" : "jbHBSFBUIAKONaiaoizfcja0f09kasndjan"
        }
      },
      "inline_response_403" : {
        "type" : "object",
        "properties" : {
          "typeError" : {
            "type" : "string",
            "example" : "ERROR_CREDENTIALS"
          },
          "messageError" : {
            "type" : "string",
            "example" : "Ha habido algún error en las credenciales introducidas"
          }
        }
      },
...
    },
...
    "requestBodies" : {
...
      "Registro" : {
        "description" : "Credenciales para el registro de nuevos usuarios/clientes",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "email", "password", "username" ],
              "type" : "object",
              "properties" : {
                "email" : {
                  "type" : "string",
                  "example" : "antonio@gmail.com"
                },
                "userName" : {
                  "type" : "string",
                  "example" : "antonio99"
                },
                "password" : {
                  "type" : "string",
                  "example" : "pass123"
                }
              }
            }
          }
        },
        "required" : true
      },
...
    }
  }
}